Movido os parsers de 2012 para o pacote ano2012
Renomeado a classe do Strategy para ParserTSEStrategy.
Adicionado as modificações no diagrama de classes.
