+++ code/ParserTSE/src/br/ufba/mata62/eleicoestransparentes/business/parser/ano2012/ComportamentoParser2012.java
@@ -23,7 +23,7 @@ public class ComportamentoParser2012 extends ComportamentoParser{
 	@Override
 	public List<Transacao> readPrestacaoContasCandidatoReceita(String uf) throws IOException {
 		String path = Path.pathRoot + Path.FLD_PRESTACAO +Path.SEPARATOR+ Path.FLD_CANDIDATO+Path.SEPARATOR+uf+Path.SEPARATOR+ Path.FILE_RECEITA_CANDIDATO;
-		return new ParserPrestacaoContasCandidatoReceita(path).parsing();
+		return new ParserPrestacaoContasCandidatoReceita2012(path).parsing();
 	}
 
 	/**
@@ -36,7 +36,7 @@ public class ComportamentoParser2012 extends ComportamentoParser{
 	@Override
 	public List<Transacao> readPrestacaoContasComiteReceita(String uf) throws IOException {
 		String path = Path.pathRoot + Path.FLD_PRESTACAO  +Path.SEPARATOR+ Path.FLD_COMITE+Path.SEPARATOR+uf+Path.SEPARATOR+  Path.FILE_RECEITA_COMITE;
-		return new ParserPrestacaoContasComiteReceita(path).parsing();
+		return new ParserPrestacaoContasComiteReceita2012(path).parsing();
 	}
 
 	/**
@@ -49,7 +49,7 @@ public class ComportamentoParser2012 extends ComportamentoParser{
 	@Override
 	public List<Transacao> readPrestacaoContasPartidoReceita(String uf) throws IOException {
 		String path = Path.pathRoot + Path.FLD_PRESTACAO  +Path.SEPARATOR+ Path.FLD_PARTIDO+Path.SEPARATOR+uf+Path.SEPARATOR+   Path.FILE_RECEITA_PARTIDO;
-		return new ParserPrestacaoContasPartidoReceita(path).parsing();
+		return new ParserPrestacaoContasPartidoReceita2012(path).parsing();
 	}
 	
 	/**
@@ -62,7 +62,7 @@ public class ComportamentoParser2012 extends ComportamentoParser{
 	@Override
 	public List<Transacao> readPrestacaoContasCandidatoDespesa(String uf) throws IOException {
 		String path = Path.pathRoot + Path.FLD_PRESTACAO +Path.SEPARATOR+ Path.FLD_CANDIDATO+ Path.SEPARATOR+uf+Path.SEPARATOR+Path.FILE_DESPESA_CANDIDATO;
-		return new ParserPrestacaoContasCandidatoDespesa(path).parsing();
+		return new ParserPrestacaoContasCandidatoDespesa2012(path).parsing();
 	}
 
 	/**
@@ -75,7 +75,7 @@ public class ComportamentoParser2012 extends ComportamentoParser{
 	@Override
 	public List<Transacao> readPrestacaoContasComiteDespesa(String uf) throws IOException {
 		String path = Path.pathRoot + Path.FLD_PRESTACAO  +Path.SEPARATOR+  Path.FLD_COMITE+Path.SEPARATOR+uf+Path.SEPARATOR+ Path.FILE_DESPESA_COMITE;
-		return new ParserPrestacaoContasComiteDespesa(path).parsing();
+		return new ParserPrestacaoContasComiteDespesa2012(path).parsing();
 	}
 
 	/**
@@ -88,7 +88,7 @@ public class ComportamentoParser2012 extends ComportamentoParser{
 	@Override
 	public List<Transacao> readPrestacaoContasPartidoDespesa(String uf) throws IOException {
 		String path = Path.pathRoot + Path.FLD_PRESTACAO  +Path.SEPARATOR+  Path.FLD_PARTIDO+Path.SEPARATOR+uf+Path.SEPARATOR+ Path.FILE_DESPESA_PARTIDO;
-		return new ParserPrestacaoContasPartidoDespesa(path).parsing();
+		return new ParserPrestacaoContasPartidoDespesa2012(path).parsing();
 	}
 	
 
@@ -102,7 +102,7 @@ public class ComportamentoParser2012 extends ComportamentoParser{
 	@Override
 	public List<Bem> readBens(String uf) throws IOException {
 		String path = Path.pathRoot +Path.FLD_BEM+Path.SEPARATOR+Path.FLD_BEM+(Path.UF_EXT.replace("$UF$", uf));
-		return new ParserBemCandidato(path).parsing();
+		return new ParserBemCandidato2012(path).parsing();
 	}
 	
 	/**
@@ -115,7 +115,7 @@ public class ComportamentoParser2012 extends ComportamentoParser{
 	@Override
 	public List<Candidato> readCandidatos(String uf) throws IOException {
 		String path = Path.pathRoot +Path.FLD_CONSULTA_CANDIDATO+Path.SEPARATOR+Path.FLD_CONSULTA_CANDIDATO+(Path.UF_EXT.replace("$UF$", uf));
-		return new ParserConsultaCandidato(path).parsing();
+		return new ParserConsultaCandidato2012(path).parsing();
 		
 	}
